+++ tests/index.test.ts
@@ -1,3 +1,4 @@
 import './access.test'
+import './noteTemplate.test'
 import './openAlexMapper.test'
 import './zoteroResponse.test'